Fix: A required popup field doesn't exist
This commit handles the case where a popup field is required but does not exist (i.e. no donation forms have been created but a form is required by the popup to create the shortcode).

There were two available solutions if required fields returned false:
1. hide the shortcode action from the dropdown
2. modify the popup body with the error and remove the primary button.

I went with the second solution as it seemed more intuitive.