Composer: fix incompatible dependencies
The `composer.lock` file has locked in a number of dependencies at version which are not compatible with the lowest PHP version the plugin is supposed to be compatible with.

By setting `config -> platform` in the `composer.json` file, the PHP version referenced there will be used for the version negotatiations, preventing these type of incompatibilities from happening in the future.