fix: properly assign associative key and value
These were meant to be the opposite (i.e. array key is the option value
and array value is the option label), but the previous implementation
mixed them up. This commit corrects this error, cleans up the code a
bit, and expands the comments to provide more clarity in regards to the
expectation of each supported registration format, as well, as comments
on how it is transformed into a `FieldOption`.