refactor: do not support associative array keys
Associative arrays are ambiguous for setting options. What should be the
label and what should be the value? Even if some decision is made, they
are limited in a way that options are not since keys must be unique and
values or labels for options do not. For this reason, we are choosing to
not support keys as labels or values. See the doc block for the method
for instruction on setting a label for an option.