feature: add ValidationRules class
Simple class for managing rules in a consistent way. Perhaps there could
be a way of validating the validation rules in the future through
providing a schema. For example, the `required` rule accepts boolean
values. This might be unnecessary since when developers implement rules
in fields or concerns, they most likely will be type hinting the API for
the rule. I believe that that would be sufficient and less complicated
than adding an additional schema for rules since it would be difficult,
nigh impossible for schema to be registered in a way that wasn’t
cumbersome to use.