chore: pin Node at 12 and npm at 7 for Volta
This is an opt-in convenience for developers. These versions will need
to be manually updated if the versions for building ever change,
however, changing these is very infrequent, while the need to switch
Node/npm version can be quite frequent for those who don’t just work on
this product or in this ecosystem.