refactor: use custom meter element
In orer to style the element in Safari, we have to set
`-webkit-appearance: none` which breaks the accessibility (i.e. the
`role` is unset) and VoiceOver ignores the element. I imagine we could
add the `role` and all the `aria-` attributes, but at that point we are
breaking a number of ARIA usage rules. This seems more reliable since we
can just set the attributes ourselves and use simpler styling. We will
just need to make sure the internal `width` is in sync with
`aria-value*` attributes (it’s not dynamic in this context so it won’t
be hard).