fix: set aria-valuenow to the raised amount
Instead of the percentage — that’s for `aria-valuetext`.

This makes a locally set CSS variable for the percent and moves the
`width` rule to the stylesheet. This is generally how I like to do
things. Colocate styles and use CSS variables to set values dynamically.

Also, formatted with PSR-12 which (surprisingly) fixed my formatting
concerns in the file.