chore: move tcpdf font copying to composer from npm
This really fits the situation better as, in terms of scope, what's
happening is inside of vendor, not node_modules. Also, running sudo on
npm scripts isn't having consistent results and gives too much sudo
power. Doing it this way only the single necessary command is getting
sudo rights to do the job.